+++ logs/log.log
@@ -1308,3 +1308,74 @@ Please let me know if there is anything else I can help you with.
 If you have any other questions or need assistance with any other topic, feel free to ask.
 2023-09-22 12:37:57,937 - INFO - Answering process completed.
 2023-09-22 12:38:49,762 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:40:14,935 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:41:32,122 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:42:09,758 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:42:27,479 - INFO - Found 778413d5-e824-59ac-8404-74ef770e9181 in Weaviate
+2023-09-22 12:42:27,576 - WARNING - Could not find 7388e4e5-0dec-57b9-b883-312bed450624 in Weaviate
+2023-09-22 12:42:28,043 - INFO - Found similar object(s) in Weaviate with similarity above threshold
+2023-09-22 12:42:28,043 - INFO - UUID: 735a0c54-3aa9-59af-81e4-a2eccdd7a8d4
+2023-09-22 12:42:28,043 - INFO - Distance: 0.043860674
+2023-09-22 12:42:28,146 - INFO - Found 9802177d-43ef-5497-8ba1-cb4eba660022 in Weaviate
+2023-09-22 12:47:17,368 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:47:38,087 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:47:46,964 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:48:06,858 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:48:07,804 - INFO - Generating output for 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:07,804 - INFO - Calling gpt-3.5-turbo with prompt: 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:08,216 - INFO - error_code=context_length_exceeded error_message="This model's maximum context length is 4097 tokens. However, your messages resulted in 7406 tokens. Please reduce the length of the messages." error_param=messages error_type=invalid_request_error message='OpenAI API error received' stream_error=False
+2023-09-22 12:48:11,789 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:48:12,545 - INFO - Generating output for 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:12,546 - INFO - Calling gpt-3.5-turbo with prompt: 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:13,233 - INFO - error_code=context_length_exceeded error_message="This model's maximum context length is 4097 tokens. However, your messages resulted in 7406 tokens. Please reduce the length of the messages." error_param=messages error_type=invalid_request_error message='OpenAI API error received' stream_error=False
+2023-09-22 12:48:21,565 - WARNING - Could not find 11c42589-3566-5f1c-81dd-3c331fe8662d in Weaviate
+2023-09-22 12:48:22,374 - INFO - Generating output for 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:22,374 - INFO - Calling gpt-3.5-turbo with prompt: 
+        Write a set of multiple-choice quiz quest...
+2023-09-22 12:48:22,651 - INFO - error_code=context_length_exceeded error_message="This model's maximum context length is 4097 tokens. However, your messages resulted in 7406 tokens. Please reduce the length of the messages." error_param=messages error_type=invalid_request_error message='OpenAI API error received' stream_error=False
+2023-09-22 12:48:25,438 - INFO - Found 85d17762-def1-5a7a-a884-4ae7870c6c00 in Weaviate
+2023-09-22 12:48:25,534 - INFO - Found 2d56dcf6-fdc3-55ff-a02e-7146e74f32bd in Weaviate
+2023-09-22 12:48:25,627 - INFO - Found a4631a7b-2847-5600-9f02-78f72e73a38c in Weaviate
+2023-09-22 12:48:45,053 - WARNING - Could not find 8b4af59b-1348-5783-b3c9-a43796141bd3 in Weaviate
+2023-09-22 12:48:45,605 - INFO - Found similar object(s) in Weaviate with similarity above threshold
+2023-09-22 12:48:45,605 - INFO - UUID: 18a17d25-c5f3-5976-b661-6f185e45f399
+2023-09-22 12:48:45,605 - INFO - Distance: 0.04655361
+2023-09-22 12:48:45,707 - WARNING - Could not find 3731cba0-e1a2-5125-9284-24b831f06bff in Weaviate
+2023-09-22 12:48:46,651 - INFO - Found similar object(s) in Weaviate with similarity above threshold
+2023-09-22 12:48:46,651 - INFO - UUID: c97d13b4-2c6e-5955-914d-bb12a2d0b4c3
+2023-09-22 12:48:46,651 - INFO - Distance: 0.038220048
+2023-09-22 12:48:46,756 - WARNING - Could not find e873c813-a3ff-5d7a-92c3-bf6565c057da in Weaviate
+2023-09-22 12:48:47,255 - INFO - Generating output for 
+    Return a glossary of key terms or jargon from...
+2023-09-22 12:48:47,255 - INFO - Calling gpt-3.5-turbo with prompt: 
+    Return a glossary of key terms or jargon from...
+2023-09-22 12:49:25,254 - INFO - Saved e873c813-a3ff-5d7a-92c3-bf6565c057da to Weaviate
+2023-09-22 12:49:37,435 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:49:47,829 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:49:48,452 - INFO - Start answering based on prompt: Tell me about LLMs and their potential for disruption.
+2023-09-22 12:49:48,452 - INFO - The top 4 chunks are considered to answer the user's query.
+2023-09-22 12:49:53,088 - INFO - The returned answer is: Based on the provided documents, I couldn't find any specific information about LLMs and their potential for disruption. It seems that the documents mainly discuss the use of LLMs, when to use them, the types of tasks they are better suited for, and how to get the most out of each prompt. 
+
+If you have any other questions or if there is anything else I can assist you with, please let me know.
+2023-09-22 12:49:53,089 - INFO - Answering process completed.
+2023-09-22 12:50:06,687 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:50:07,268 - INFO - Start answering based on prompt: How do I get the most out of LLM prompts?.
+2023-09-22 12:50:07,268 - INFO - The top 4 chunks are considered to answer the user's query.
+2023-09-22 12:50:09,950 - INFO - The returned answer is: Based on the provided examples, the answer to the question "How do I get the most out of LLM prompts?" is not found in the documentation. Please rephrase your query with more details or let me know how else I can assist you.
+2023-09-22 12:50:09,950 - INFO - Answering process completed.
+2023-09-22 12:50:24,947 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:50:34,720 - INFO - Anonymized telemetry enabled. See https://docs.trychroma.com/telemetry for more information.
+2023-09-22 12:50:35,363 - INFO - Start answering based on prompt: What does the text say about sprint planning?.
+2023-09-22 12:50:35,363 - INFO - The top 4 chunks are considered to answer the user's query.
+2023-09-22 12:50:38,881 - INFO - The returned answer is: Based on the provided examples, the text says the following about sprint planning:
+
+- LLMs should be able to give us a good idea about sprint planning.
+- The soft skills needed to prioritize and push back are sadly not present in ChatGPT.
+
+If you have any other questions or need assistance with anything else, feel free to ask!
+2023-09-22 12:50:38,881 - INFO - Answering process completed.
