+++ requirements.txt
@@ -7,7 +7,6 @@ tiktoken
 langchain
 pypdf
 weaviate-client
-streamlit-chat
 docx2txt
 pysqlite3-binary
 chromadb
